Split the react components out into separate files
- I'm using one folder per component, with child components nested under the parent folder.
- refs #18
