+++ web/app/package.json
@@ -12,5 +12,8 @@
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
+  },
+  "devDependencies": {
+    "@types/react": "^16.4.11"
   }
 }
