--- web/app/src/App.css
@@ -1,45 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-header {
-  background-color: #222;
-  height: 50px;
-  padding: 20px;
-  color: white;
-}
-
-.App-title {
-  font-size: 1.5em;
-}
-
-.App-intro {
-  font-size: large;
-}
-
-@keyframes App-logo-spin {
-  from { transform: rotate(0deg); }
-  to { transform: rotate(360deg); }
-}
-
-/* ------------------- */
-
-.Story {
-  text-align: center;
-}
-
-.Story-header {
-  background-color: #222;
-  height: 100px;
-  padding: 20px;
-  color: white;
-}
-
-.Story-title {
-  font-size: 1.5em;
-  font-weight: bold;
-}
-
-.Story-author {
-  color: #ccc;
-}
