+++ web/app/src/Story/index.css
@@ -1,0 +1,20 @@
+
+.Story {
+  text-align: center;
+}
+
+.Story-header {
+  background-color: #222;
+  height: 100px;
+  padding: 20px;
+  color: white;
+}
+
+.Story-title {
+  font-size: 1.5em;
+  font-weight: bold;
+}
+
+.Story-author {
+  color: #ccc;
+}
